finish 7_chkp

+++ app/controllers/bookmarks_controller.rb
@@ -47,18 +47,27 @@ class BookmarksController < ApplicationController
     end
     
   def destroy
-      @bookmark = Bookmark.find(params[:id])
-      puts '-----------------'
-      puts params
-      puts '==========================='
- 
-      #authorize @bookmark
+     @bookmark = Bookmark.find(params[:id])
+
+
+     authorize @bookmark
+     
      if @bookmark.destroy
-       flash[:notice] = "\"#{@bookmark.url}\" was deleted successfully."
-       redirect_to topic_path(params[:topic_id])
+      flash[:notice] = "\"#{@bookmark.url}\" was deleted successfully."
+      redirect_to topic_path(params[:topic_id])
      else
-       flash.now[:alert] = "There was an error deleting the bookmark."
-       redirect_to topic_path(params[:topic_id])
+      flash.now[:alert] = "There was an error deleting the bookmark."
+      redirect_to topic_path(params[:topic_id])
      end
   end
+  
+  def user_not_authorized
+      puts '------------------------------'
+      puts params
+      puts '+++++++++++++++++++++++++++++++++'
+    flash[:alert] = "You are not cool enough to do this - go back from whence you came."
+    redirect_to(topic_path(params[:topic_id]))
+  end
+  
+
 end
